Move ConstantSource to separate file
