+++ volumina/pixelpipeline/datasources/__init__.py
@@ -34,100 +34,14 @@ from volumina.slicingtools import is_pure_slicing, slicing2shape, is_bounded, ma
 from volumina.config import CONFIG
 import numpy as np
 from future.utils import raise_with_traceback
+
 from .array import ArraySource, ArraySinkSource, RelabelingArraySource
+from .constant import ConstantSource
 
 try:
     from .lazyflow import LazyflowSource, LazyflowSinkSource
 except ImportError:
     pass
-
-
-# *******************************************************************************
-# C o n s t a n t R e q u e s t                                                *
-# *******************************************************************************
-
-
-class ConstantRequest(object):
-    def __init__(self, result):
-        self._result = result
-
-    def wait(self):
-        return self._result
-
-    def getResult(self):
-        return self._result
-
-    def cancel(self):
-        pass
-
-    def submit(self):
-        pass
-
-    def adjustPriority(self, delta):
-        pass
-
-
-assert issubclass(ConstantRequest, RequestABC)
-
-# *******************************************************************************
-# C o n s t a n t S o u r c e                                                  *
-# *******************************************************************************
-
-
-class ConstantSource(QObject):
-    isDirty = pyqtSignal(object)
-    idChanged = pyqtSignal(object, object)  # old, new
-    numberOfChannelsChanged = pyqtSignal(int)  # Never emitted
-
-    @property
-    def constant(self):
-        return self._constant
-
-    @property
-    def numberOfChannels(self):
-        return 1
-
-    @constant.setter
-    def constant(self, value):
-        self._constant = value
-        self.setDirty(sl[:, :, :, :, :])
-
-    def __init__(self, constant=0, dtype=np.uint8, parent=None):
-        super(ConstantSource, self).__init__(parent=parent)
-        self._constant = constant
-        self._dtype = dtype
-
-    def clean_up(self):
-        pass
-
-    def id(self):
-        return id(self)
-
-    def request(self, slicing, through=None):
-        assert is_pure_slicing(slicing)
-        assert is_bounded(slicing)
-        shape = slicing2shape(slicing)
-        result = np.full(shape, self._constant, dtype=self._dtype)
-        return ConstantRequest(result)
-
-    def setDirty(self, slicing):
-        if not is_pure_slicing(slicing):
-            raise Exception("dirty region: slicing is not pure")
-        self.isDirty.emit(slicing)
-
-    def __eq__(self, other):
-        if other is None:
-            return False
-        return self._constant == other._constant
-
-    def __ne__(self, other):
-        return not (self == other)
-
-    def dtype(self):
-        return self._dtype
-
-
-assert issubclass(ConstantSource, SourceABC)
 
 
 class MinMaxUpdateRequest(object):
